+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 1.1.0 - 2020-01-12
+- Removed `\RQuadling\Console\Input\Input` as not really appropriate for this library.
+  Instead, this library will use whatever `\Symfony\Component\Console\Input\InputInterface` and 
+  `\Symfony\Component\Console\Output\OutputInterface` compliant objects are supplied.
+
 ## 1.0.0 - 2019-12-27 - Initial release